requirements.txt
python-dotenv
cryptography
tk